pyproject.toml
[tool.poetry]
name = "iot-master"
version = "0.1.0"
description = ""
authors = ["ewasince <ewasince@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11.1"

[tool.poetry.group.dev.dependencies]
black = "^23.10.0"

[tool.poetry.group.simple_console_light.dependencies]
flask = "^3.0.0"

[tool.poetry.group.tg_bot.dependencies]
aiogram = "^3"
voice-assistant = {git = "https://github.com/Ewasince/voice_assistant"}
soundfile = "^0.12.1"
pyyaml = "^6.0.1"
numpy = "^1.26.1"
#pymystem3 = "^0.2.0"
#pymorphy3 = "^1.2.1"

[tool.poetry.group.homassistant.dependencies]
homeassistant = "^2023.9.3"
websockets = "^12.0"
voice-assistant = {git = "https://github.com/Ewasince/voice_assistant"}
urllib3 = "==1.26.16"

[build-system]
requires = ["poetry-core>=1.6"]
build-backend = "poetry.core.masonry.api"
